+++ doc/user/content/release-notes.md
@@ -140,6 +140,10 @@ List new features before bug fixes.
 - Fix a crash when decoding certain messages from Protobuf-formatted Kafka
   topics {{% gh 8930 %}}.
 
+- Improve support for table functions in `SELECT` lists.
+
+- Support PgJDBC's `getPrimaryKeys()` API.
+
 {{% version-header v0.10.0 %}}
 
 - Allow ingesting avro schemas whose top node is not a record type.